requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.1
watchdog==4.0.1
